Fix unit test failing on certain machines
Ubuntu 14.04 only seems to be able to seek to an offset of 2^32 blocks or
2^32*4096 bytes.

+++ source/io/file/stream.d
@@ -626,8 +626,8 @@ struct FileBase
         assert(f.skip(5) == 10);
         assert(f.seekTo(-5, From.end) == data.length - 5);
 
-        // Test large offset
-        assert(f.seekTo(long.max) == long.max);
+        // Test offset beyond the end of the file
+        assert(f.seekTo(int.max) == int.max);
     }
 
     /**
